Support for VSM. (#34)
Under Windows VSM reading memory can actually produce an error. We
therefore need to be able to trap the error and mark the page as
unreadable in the AFF4 map properly.

+++ aff4/aff4_image.h
@@ -160,7 +160,7 @@ class AFF4Image: public AFF4Stream {
         AFF4Stream* source,
         ProgressContext* progress = nullptr) override;
 
-    AFF4Status Write(const char* data, int length) override;
+    AFF4Status Write(const char* data, size_t length) override;
 
     /**
      * Read data from the current read pointer.
